requirements.txt
sounder
pyaudio
PyDispatcher
eyeD3
pygame
gTTS
SpeechRecognition
Metaphone
newsapi
soccerpy
pyball
requests
evernote3
pytz
facebook-sdk
dateutil
Httplib2
apiclient
oauth2client
omdb
psutil
tweepy
Pyowm
webbrowser
urllib3
